Add playback and keyboard hooks
